# api-test/src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"SceneResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"deviceId": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"deviceId",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateSceneRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"deviceId": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"deviceId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateSceneRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"deviceId": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"DeviceResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"ip": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"ip",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateDeviceRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"ip": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"ip"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateDeviceRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"ip": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "api-test",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/scenes": {
			"get": {
				"operationId": "GetScenesByDevice",
				"responses": {
					"200": {
						"description": "Retrieved scenes successfully",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SceneResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all scenes for a specific device",
				"tags": [
					"Scenes"
				],
				"security": [],
				"parameters": [
					{
						"description": "The device ID to get scenes for",
						"in": "query",
						"name": "deviceId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateScene",
				"responses": {
					"201": {
						"description": "Scene created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SceneResponse"
								}
							}
						}
					}
				},
				"description": "Create a new scene",
				"tags": [
					"Scenes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateSceneRequest"
							}
						}
					}
				}
			}
		},
		"/scenes/{sceneId}": {
			"put": {
				"operationId": "UpdateScene",
				"responses": {
					"200": {
						"description": "Scene updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SceneResponse"
								}
							}
						}
					}
				},
				"description": "Update a scene by ID",
				"tags": [
					"Scenes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "sceneId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateSceneRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteScene",
				"responses": {
					"200": {
						"description": "Scene deleted successfully",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Delete a scene by ID",
				"tags": [
					"Scenes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "sceneId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/devices": {
			"get": {
				"operationId": "GetDevices",
				"responses": {
					"200": {
						"description": "Retrieved devices successfully",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/DeviceResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all devices",
				"tags": [
					"Devices"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateDevice",
				"responses": {
					"201": {
						"description": "Device created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeviceResponse"
								}
							}
						}
					}
				},
				"description": "Add a new device",
				"tags": [
					"Devices"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateDeviceRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteAllDevices",
				"responses": {
					"200": {
						"description": "All devices and scenes deleted successfully",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"deletedScenes": {
											"type": "number",
											"format": "double"
										},
										"deletedDevices": {
											"type": "number",
											"format": "double"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"deletedScenes",
										"deletedDevices",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Delete all devices and their related scenes",
				"tags": [
					"Devices"
				],
				"security": [],
				"parameters": []
			}
		},
		"/devices/{deviceId}": {
			"put": {
				"operationId": "UpdateDevice",
				"responses": {
					"200": {
						"description": "Device updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeviceResponse"
								}
							}
						}
					}
				},
				"description": "Update a device by ID",
				"tags": [
					"Devices"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "deviceId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateDeviceRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteDevice",
				"responses": {
					"200": {
						"description": "Device deleted successfully",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Delete a device by ID",
				"tags": [
					"Devices"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "deviceId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}